+++ packages/e2e/android/src/main/java/dev/flutter/plugins/e2e/E2EPlugin.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package dev.flutter.plugins.instrumentationadapter;
+package dev.flutter.plugins.e2e;
 
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
@@ -12,8 +12,8 @@ import io.flutter.plugin.common.PluginRegistry.Registrar;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
-/** InstrumentationAdapterPlugin */
-public class InstrumentationAdapterPlugin implements MethodCallHandler {
+/** E2EPlugin */
+public class E2EPlugin implements MethodCallHandler {
 
   public static CompletableFuture<Map<String, String>> testResults = new CompletableFuture<>();
